demo/posix/cyclictest: Fix building with glibc >= 2.41
glibc 2.41 introduces wrappers for sched_setattr() and sched_getattr()
with slightly different signatures, so gcc (v15) complained about
redefinitions.

We can remove the prototypes from Xenomai as we never use them.
struct sched_attr was only used by the prototypes, so it can be removed as well.

The only remaining piece in rt-sched.h would be the definition of
SCHED_DEADLINE, which is part of glibc since 2.26. For the Debian
world this predates Debian buster, which is currently the oldest
supported version.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>