lib/cobalt: Fix cancellation issues over pthread_cond_[timed]wait
First of all, the usage of cobalt_cond_cleanup_t::count got broken long
ago, leaving this field initialized to 0 rather than the actual state on
entry. This could have corrupted lockcnt of the associated mutex when
triggering the cleanup.

The second issue, and that was always broken, is that cancellation was
only checked outside of the -EINTR look. But as cancellation actually
triggers an early return from the cobalt syscall with -EINTR, we never
canceled and just continued to block unless async cancellation was
enabled.

Fixes: a90f68351124 ("cobalt: more mutex and condvars cleanups")
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>