drivers/net: fix compilation error caused by RTCFG_DEBUG
When kernel config enable CONFIG_DYNAMIC_DEBUG and
CONFIG_XENO_DRIVERS_NET_RTCFG_DEBUG
the RTCFG_DEBUG macro uses a ternary operator (?:) to
conditionally call pr_debug based on the debug level
(rtcfg_debug >= n). The pr_debug macro, in turn, expands to
dynamic_pr_debug, which is defined in include/linux/dynamic_debug.h
and involves a do { ... } while (0) construct. The problem arises
because the ternary operator's false branch returns 0, which is
not a valid expression in the context of the do block, leading to
the error expected expression before ‘do’.

Signed-off-by: liuyang <liuyang@sinsegye.com.cn>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>