lib/cobalt: arm: Skip SIGSHADOW_ACTION_BACKTRACE over asynchronous signals
The unwinders of libgcc and libunwind prefer the ARM exception handling
table over debug information. Unwinding this way is not designed to
handle asynchronous signals, causing wrong backtraces in the best case
but also too often even crashes.

Avoid running into that trap by skipping the request to create a
backtrace via SIGSHADOW on ARM in case either the migration reason was a
signal, or PTHREAD_WARNSW is enabled for the current thread. In both
cases, the SIGSHADOW could be raised over the first instruction of
another signal handler, causing the unwinder to fail.

Improving the unwinding libraries themselves is likely no longer worth
the effort for this phase-out architecture.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>