docs(xeno3): change to client-side search
The Xenomai3 documentation is hosted on gitlab pages which publishes
static websites. With the option SERVER_BASED_SEARCH set to YES, it
tries to tun a php script on the server side, which just leads in this
setup to a download prompt to be shown. By turning this option off (and
keeping SEARCHENGINE on YES), the search will be performed on the client
side. This makes it at least possible to find symbols and have
livesearch enabled. However, it does not allow for full-text search.

More information can be found here:
https://www.doxygen.nl/manual/searching.html

Signed-off-by: Markus Kurz <m.kurz92@gmail.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>