docs(xeno3): adjust the size of the search bar
The search bar used to search through the API documentation had a
font-size of 8px, which was barely visible. Therefore, this commit
adjusts the font-size of the search bar, the drop down menu of the
search, and also the livepreview to 1.2em of the base font-size of the
documentation which is 14px.

Changes to the xeno3prm-html.conf.in had to be made in order for doxygen
to use the extra stylesheet.

Signed-off-by: Markus Kurz <m.kurz92@gmail.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>