y2038: lib/cobalt: select: Normalize timeout value in the y2038 case
Andreas discovered a difference in the timeout value handling between
the native and compat + time64_t (= y2038 support) cases.

The time64_t based select() service is internally routed to the
pselect() service. Linux does a normalization step for the timeout
values of select() and so does cobalt. Both, Linux and cobalt skip this
step for the pselect() syscall. That means that we were missing one
normalization step for select() in the y2038 case.

Reported-by: Andreas Ziegler <ziegler.andreas@siemens.com>
Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>