alchemy/task: Initialize thread handle on rt_task_bind
This field was so far left alone, risking troubles if a function
evaluating it using a stale value. That's so far rt_task_join, but we
will soon use it also for additional integrity validation.

So, first initialize it to 0, on Linux an invalid pthread_t value. Then,
if the task is a local one, pick up the ptid from the tcb so that even
rt_task_join would start to work for process-local tasks.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>