alchemy/task: Match thread handle on task lookup as well
This can catch stale task objects, namely when the original thread
already terminated, its task handle was finalized and dropped from the
registry but then reused by a new task object. The new one would
normally get a new thread handle as well, allowing us to detect the
stale of the original task object.

We just need to take into account the remote tasks or tasks currently in
the process of being bound do not have a valid thread handle.

Suggested-by: Mauro S <mau.salvi@yahoo.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>